s32v234: Corect clock frequency formula
Currently we loose precision by casting to integer partial results
instead of casting only the final result.

Issue: ALB-2085

+++ arch/arm/cpu/armv8/s32/s32v234/clock.c
@@ -1,6 +1,6 @@
 /*
  * (C) Copyright 2015-2016 Freescale Semiconductor, Inc.
- * (C) Copyright 2017 NXP
+ * (C) Copyright 2017-2018 NXP
  *
  * SPDX-License-Identifier:	GPL-2.0+
  */
@@ -98,7 +98,8 @@ static int program_pll( enum pll_type pll, u32 refclk_freq, u32 freq0, u32 freq1
 	/*
 	 * This formula is from platform reference manual (Rev. 1, 6/2015), PLLDIG chapter.
 	 */
-	fvco = (refclk_freq / plldv_prediv) * (plldv_mfd + pllfd_mfn/(float)20480);
+	fvco = (refclk_freq / (float)plldv_prediv) *
+		(plldv_mfd + pllfd_mfn/(float)20480);
 
 	/*
 	 * VCO should have value in [ PLL_MIN_FREQ, PLL_MAX_FREQ ]. Please consult
